Move db.models to models
